% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedarray.R
\name{setup_nametable}
\alias{setup_nametable}
\title{Set up an options/parameters system}
\usage{
setup_nametable(argnames, argranges)
}
\arguments{
\item{argnames}{a character vector of selector names}

\item{argranges}{a list of character vectors of valid selector valuess. Each selector is the name of a dimension of the (internal) array}
}
\value{
a named list of 2 functions : set and get
}
\description{
This function initializes an aray of options/parameters (of any
type), accessible through a combination of selector names.
It returns a named list of set/get function, which respectively
set/get the values stored in an internal array of values
}
\details{
The `setup_nametable` function returns a list of
2 functions:
\itemize{
\item \code{get} returns the values of the selected  array fields, and `
\item \code{set} assigns the given \code{value} to the selected fields of the aray.
}
Both functions return the selected array values
}
\examples{
# setup
tst2 <- setup_nametable(argnames = c("bla", "ble", "bli"),
                        argranges = list(c("a", "b"),
                                         c("u", "v", "w"),
                                         c("x", "y", "z", "t")))

# setting and retrieving value
tst2$set( value = LETTERS[1:24])
tst2$get( bla = "a", ble = "u", bli = "z")
tst2$get( bla = "a", ble = "u")
tst2$set( bla = "a", ble = "u", value = "ZZ")
tst2$get()
}
\author{
Bruno Fischer Colonimos
}
